updated dependencies

+++ app/Model/Taxes.php
@@ -69,7 +69,7 @@ final class Taxes {
         $return->denizens[$d->id] = $d;
         continue;
       }
-      $d->income = (int) array_sum($this->calculateIncome($denizen->id, $month ?? (int) date("n"), $year ?? (int) date("Y")));
+      $d->income = array_sum($this->calculateIncome($denizen->id, $month ?? (int) date("n"), $year ?? (int) date("Y")));
       $d->tax = $this->calculateTax($d->income);
       $return->denizens[$d->id] = $d;
       $return->taxes += $d->tax;
